+++ envoy-control-runner/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/snapshot/debug/SnapshotDebugInfo.kt
@@ -61,3 +61,7 @@ data class SnapshotDebugInfo(
         versions = Versions()
     )
 }
+
+data class EndpointInfoList(val clusters: List<EndpointInfo> = emptyList())
+
+data class EndpointInfo(val datacenter: String, val ip: String, val port: Int)