+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/config/envoy/EnvoyContainer.kt
@@ -127,4 +127,6 @@ class EnvoyContainer(
     fun adminUrl() = "http://$containerIpAddress:${getMappedPort(ADMIN_PORT)}"
 
     fun admin() = EnvoyAdmin(adminUrl())
+
+    fun ingressHost(): String = ingressListenerUrl().removePrefix("http://")
 }